Updates test scripts

+++ testfiles/do-riboaligner-parallel-tests.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
-set -e
+
 $RIBODIR/ribotest.pl -f $RIBODIR/testfiles/riboaligner.p.testin ra-p-test
-rm -rf ra-p-test
+if [ $? == 0 ]; then
+   rm -rf ra-p-test
+   echo "Success: all tests passed"
+   exit 0
+else 
+   echo "FAIL: at least one test failed"
+   exit 1
+fi
